Expand item stacks, fix armor enchanting, and refine Twilight assets
Support 9999 potions and 999 enchant cards across inventory and storage. Remove non-enchant GM cards, thin and texture Twilight blade edges, add ten weapon concepts, and capture local client crash diagnostics. Record live validation and remaining crash investigation.

+++ src/client/gameclient/Src/Game.cpp
@@ -2362,7 +2362,7 @@ void CGame::OnEventBuyItem( DWORD dwParam1, DWORD dwParam2 )
 {
 	int			nItemType	= dwParam1 >> 16;		// 구입할 아이템
 	WORD		wShop		= dwParam1 & 0xFF;		// SHOP
-	BYTE		btCnt		= dwParam2;				// 수량(포션용) from sk8snow
+	WORD		btCnt		= dwParam2;				// 수량(포션용) from sk8snow
 
 	CPacket		sendPacket;
 	sendPacket.Tag( dPACKET_BUY_ITEM )
